ogma-cli: Make FPrime backend receive options as record. Refs #204.
The FPrime backend has a large amount of input options, making it
cumbersome to pass them around.

A prior commit has modified the backend in ogma-core to packs those
options in a record, making the signature of the top-level function
simpler and more uniform.

This commit adapts the CLI to provide the arguments in a record as
required by the updated backend.