ogma-core: Make FPrime backend able to handle different formats. Refs #204.
The standalone backend supports JSON and XML, which can be further
confirured in the command line, but that capability is not yet available
in the FPrime backend.

This commit adds that capability by introducing the capability of
handling JSON and XML file formats, and reading the specific paths
of the elements needed in properties from the command line.