ogma-core: Extend FPrime backend to use external process to parse properties. Refs #204.
There is a need in Ogma to make the frontend more versatile, and to
extend it to support languages unknown to Ogma without users having to
modify the code of Ogma itself. To that end, we want to empower users
with the ability to use a custom command to transform individual
properties into a format that Ogma understands, and have Ogma call that
external command on demand. This capability is already available in the
standalone backend, but not in other backends.

This commit modifies the FPrime backend to take an auxiliary command that
is used as the process name to call to parse individual properties.