ogma-language-xlsx: Introduce library to read specs from XLSX files. Refs #263.
To facilitate importing requirements and properties from different
tools, there is a need to extend Ogma to process XLSX files, which are
frequently available as an export option in many MBSE tools and used in
many real-world projects.

This commit introduces a new library that allows reading properties or
requirements from XLSX files.