ogma-core: Introduce overview command. Refs #272.
As Ogma adds support for MBSE, it is useful to add an overview or
summary command that shows a list of all the information that Ogma knows
about the files loaded.

This command adds a new Overview module to `ogma-core`, which summarizes
the information of the files loaded and can be used by multiple
frontends (e.g., GUI, CLI), and exposes the command in the Cabal file.