ogma-cli: Expose overview command. Refs #272.
As Ogma adds support for MBSE, it is useful to add an overview or
summary command that shows a list of all the information that Ogma knows
about the files loaded.

A prior commit has added an overview command to `ogma-core` that
summarizes the information read from the input files.

This commit exposes that command via the command-line interface.