ogma-cli: Add tested file to tests directory. Refs #216.
The Cabal file for ogma-cli is not including all files needed for the
tests, or declaring them in the Cabal package, causing the tests to fail
when the packages are unpacked from the distributable source packages
created with Cabal.

This commit adds a missing file to the tests/ directory, which is its
natural location, and makes it possible for the tests access the file
without having to use getDataDir. The file is added under a new name for
consistency with other files used by the tests.