ogma-cli: Change location of test file in test suite. Refs #216.
The Cabal file for ogma-cli is not including all files needed for the
tests, or declaring them in the Cabal package, causing the tests to fail
when the packages are unpacked from the distributable source packages
created with Cabal.

A prior commit has added a missing example to the tests/ directory.

This commit changes the tests to obtain that auxiliary test file from
said directory. It also adjusts the file name to its new name under the
tests/ directory.