ogma-cli: List all auxiliary test files in Cabal package. Refs #216.
The Cabal file for ogma-cli is not including all files needed for the
tests, or declaring them in the Cabal package, causing the tests to fail
when the packages are unpacked from the distributable source packages
created with Cabal.

Prior commits have added any missing auxiliary files to the tests/
directory, and adjusted the tests accordingly to find them.

This commit changes the Cabal package to include all auxiliary test
files to the extra-source-files field, making sure that the files are
included in the distributable source package but not installed with the
binary distribution.