ogma-language-cocospec: Remove references to old design of Ogma from documentation. Refs #220.
Some of the documentation of Ogma, such as the READMEs of the CoCoSpec
and smv language processing modules, still contains references to the
way that Ogma was designed in older version. As Ogma has evolved over
time and requirements have changed, these references are no longer
applicable.

This commit removes references to Ogma's old design from the
documentation.