ogma-language-c: Upgrade version bounds of dependencies on Cabal. Refs #213.
In order to keep Ogma effectively working in the current Haskell ecosystem, as
well as new versions of GHC, we need to extend the versions of dependencies
that Ogma can be installed with. More specifically, we currently impose a
restriction on Cabal to be lower than version 3.9, which is preventing Ogma
from being used in Debian sid.

This commit modifies version bounds of ogma-language-c to allow installation
with Cabal 3.14, the most recent version of Cabal that does not require a
change to the Haskell code.