ogma-extra: Bump upper version constraint on text. Refs #225.
In order to keep Ogma effectively working in the current Haskell ecosystem, and
with new versions of GHC, we want to extend the versions of dependencies that
Ogma can be installed with as much as possible to include newly released
versions. The current constraints prevent installation of Ogma with GHC 9.8 and
GHC 9.10, and they will prevent inclusion in Stackage.

This commit relaxes the upper bounds on the version constraints on text so
that Ogma can be installed with versions up (but not including) to 2.2.