ogma-extra: Remove extraneous EOL character. Refs #224.
Several files that include the license header include an extraneous unicode-16
character representing a line separator (hex code 0x20 0x28). This character
doesn't belong, it doesn't always render well and it's just a result of copy
and paste from another document into the code. It should be removed.

This commit removes all occurrences of said character in the library
ogma-extra.