ogma-language-c: Specify tools needed using build-tool-depends. Refs #149.
The Cabal packages in Ogma were written for an older version of Cabal.
Currently, they only state a dependency on `BNFC` as part of a custom-setup
section. However, this is sub-ideal because `alex` and `happy` are also needed,
and Cabal fails to detect that requirement, and also fails to install `BNFC`
and make it available before it tries to compile Ogma. This makes the
installation instructions overly complicated (all three of those packages need
to be installed by hand first), and makes Hackage report that Ogma as a whole
cannot be built.

This commit updates the Cabal package to specify the tools it needs using
`build-tool-depends`. The notation is used by versions of Cabal from 3.0 to
modern versions when using the standard `install` command, which ensures that
the programs needed are installed before. When using `v1-install`, which is
accepted by versions of Cabal as far as 2.4, the programs `alex`, `happy` and
`bnfc` need to be installed prior to installing Ogma (like before).