ogma-cli: Add cabal.project. Refs #149.
The Cabal packages in Ogma were written for an older version of Cabal.
Currently, they only state a dependency on BNFC as part of a custom-setup
section. However, this is sub-ideal because alex and happy are also needed, and
Cabal fails to detect that requirement, and also fails to install BNFC and make
it available before it tries to compile Ogma. This makes the installation
instructions overly complicated (all three of those packages need to be
installed by hand first), and makes Hackage report that Ogma as a whole cannot
be built.

Prior commits have updated the Cabal files to use build-tool-depends to
indicate which tools need to be available before compiling some packages.

This commit adds a cabal.project file at the top-level so that newer versions
of Cabal find all Ogma packages in the current directory, which is needed to be
able to exploit that functionality.