ogma-core: Make structured data available to cFS template. Refs #229.
The current cFS backend is overly specific, complex and composes data manually.
For example, it assumes a specific structure for the message handlers. It also
extracts data from data structures, manually converts it so strings, and then
packs it into a JSON structure. This limits user's abilities to adjust the
template to fit their needs, since most of the data is given as fixed strings
(e.g., variable declarations are given as multiple lines with the shape
"<variable_type> <variable_name>;", rather than letting users access each
variable declaration and, within them, each type and name individually.

This commit simplifies the cFS backend by packing all data that may be relevant
for the cFS application in a data structure, converting it to JSON, and then
passing it to the template expansion system.