ogma-core: Update template to use new variable names and structures. Refs #229.
The current cFS backend is overly specific, complex and composes data manually.
For example, it assumes a specific structure for the message handlers. It also
extracts data from data structures, manually converts it so strings, and then
packs it into a JSON structure. This limits user's abilities to adjust the
template to fit their needs, since most of the data is given as fixed strings
(e.g., variable declarations are given as multiple lines with the shape
"<variable_type> <variable_name>;", rather than letting users access each
variable declaration and, within them, each type and name individually.

A prior commit has modified the cFS backend to use provide more raw data to the
cFS template, making it more versatile.

This commit updates the template to make use of such data, by moving some of
the strings that the prior cFS backend was composing together into the template
directly. The template is modified such that it generates the exact same final
cFS application as before; this makes the template specific to ICAROUS, which
is preferable to having the cFS backend be specific to ICAROUS.