ogma-core: Update Copilot struct code generator to use new function names. Refs #231.
Copilot 3.19 changed the way that structs are generated. For example, the
function `typename` was removed in favor of `typeName`. The code currently
generated by Ogma requires Copilot < 3.19 to work.

This commit updates the code generator of Copilot structs to use `typeName`
instead. That is the only function affected by that change in Copilot that Ogma
currently uses in this backend.