ogma-core: Update cFS backend to process a handlers file. Refs #234.
The cFS backend currently generates code that would normally have to be
adjusted by the user. The default cFS template assumes that there's only one
error handler, and it's name is fixed to split. This is clearly sub-ideal: we
want users to be able to generate code that they do not have to modify.

This commit updates `ogma-core` so that the cFS backend accepts an additional
parameter with a file listing handlers. The file, if available, is opened and
the handler names are made available to the template.