ogma-core: Update template to declare trigger handlers. Refs #234.
The cFS backend currently generates code that would normally have to be
adjusted by the user. The default cFS template assumes that there's only one
error handler, and it's name is fixed to split. This is clearly sub-ideal: we
want users to be able to generate code that they do not have to modify.

A prior commit has updated `ogma-core` to accept a handlers file and, when
available, pass the handler names listed in it to the template.

This commit updates the template to make use of those trigger names, and
publish a notification in the form of a cFS event for each one of them.