ogma-cli: Expose handlers-file argument to cFS backend in CLI. Refs #234.
The cFS backend currently generates code that would normally have to be
adjusted by the user. The default cFS template assumes that there's only one
error handler, and it's name is fixed to split. This is clearly sub-ideal: we
want users to be able to generate code that they do not have to modify.

Prior commits have updated `ogma-core` to accept a handlers file and, when
available, populate the corresponding handler declarations in the cFS
application code.

This commit exposes that additional argument to the user via an optional
argument to the `cfs` command in the CLI.