ogma-cli: Update README to document new template variable. Refs #234.
The cFS backend currently generates code that would normally have to be
adjusted by the user. The default cFS template assumes that there's only one
error handler, and it's name is fixed to split. This is clearly sub-ideal: we
want users to be able to generate code that they do not have to modify.

Prior commits have updated `ogma-cli` to accept a handlers file and, when
available, populate the corresponding handler declarations in the cFS
application code.

This commit documents the new argument to the `cfs` command, and the variable
used in cFS application templates to hold the names of the handlers or fault
triggers used by the monitoring application. One of the limitations listed in
the README, related to the lack of flexibility in the cFS backend, no longer
applies, so it is now removed.