ogma-cli: Add new example file with a list of handlers. Refs #234.
The cFS backend currently generates code that would normally have to be
adjusted by the user. The default cFS template assumes that there's only one
error handler, and it's name is fixed to split. This is clearly sub-ideal: we
want users to be able to generate code that they do not have to modify.

Prior commits have updated `ogma-cli` to accept a handlers file and, when
available, populate the corresponding handler declarations in the cFS
application code.

This commit adds an additional example file to the `examples/` directory, which
can be used by users to generate a cFS monitoring application.