ogma-core: Update cFS backend to process a template variables file. Refs #106.
The generated cFS application may need definitions that come from external
files, such as message IDs from other applications, or structs defined
elsewhere. Users may also want to customize additional aspects of the generated
application without having to modify the template, such as passing additional
variables to be expanded in the template.

This commit extends the cFS backend to expect an additional argument denoting a
JSON file with an object whose keys will be expanded as variables in the cFS
template.