ogma-cli: Expose template vars argument to cFS backend in CLI. Refs #106.
The generated cFS application may need definitions that come from external
files, such as message IDs from other applications, or structs defined
elsewhere. Users may also want to customize additional aspects of the generated
application without having to modify the template, such as passing additional
variables to be expanded in the template.

A prior commit has extended the cFS backend to expect an additional argument
filename containing a JSON object, whose keys are expanded as variables in the
cFS template.

This commit exposes the argument to the user in the command-line via an
optional CLI argument.