ogma-cli: Update README to document new template vars argument. Refs #106.
The generated cFS application may need definitions that come from external
files, such as message IDs from other applications, or structs defined
elsewhere. Users may also want to customize additional aspects of the generated
application without having to modify the template, such as passing additional
variables to be expanded in the template.

Prior commit have extended the `cfs` command to accept an optional argument
with a filename containing a JSON object, whose keys are expanded as variables
in the cFS template.

This commit documents the new argument to the `cfs` command, and how variables
are expanded in cFS application templates.