ogma-core: Replace ICAROUS-specific function declaration in template. Refs #237.
The cFS application template was originally built for ICAROUS, but it is now
generic to cFS. ICAROUS is still mentioned in the template, and the resulting
code will not compile unless ICAROUS is present as an app.

This commit replaces a declaration of an ICAROUS-specific message processing
method in a header file in the cFS template, which presumably will not exist if
ICAROUS is not publishing messages that the monitoring app consumes, with a
generic version that generates the specific function names to declare based on
variables already available to the template.