ogma-core: Replace ICAROUS-specific includes in template. Refs #237.
The cFS application template was originally built for ICAROUS, but it is now
generic to cFS. ICAROUS is still mentioned in the template, and the resulting
code will not compile unless ICAROUS is present as an app.

This commit removes ICAROUS-specific includes in the cFS application
implementation in the template with a generic expansion based on a new JSON
variable.

Presumably, additions to the header of the implementation will be given in
separate lines. To aid readability, the variable given should be an array
instead of a plain string.