ogma-core: Replace ICAROUS-specific comment in template. Refs #237.
The cFS application template was originally built for ICAROUS, but it is now
generic to cFS. ICAROUS is still mentioned in the template, and the resulting
code will not compile unless ICAROUS is present as an app.

This commit modifies a comment that specifically mentions ICAROUS data, to
refer to generic cFS data.