ogma-core: Replace ICAROUS-specific include in template. Refs #237.
The cFS application template was originally built for ICAROUS, but it is now
generic to cFS. ICAROUS is still mentioned in the template, and the resulting
code will not compile unless ICAROUS is present as an app.

This commit replaces an ICAROUS-specific include in the compilation scripts for
the monitoring app with a new array template variable that contains the list of
dependencies.