ogma-cli: Document new template variables in README. Refs #237.
The cFS application template was originally built for ICAROUS, but it is now
generic to cFS. ICAROUS is still mentioned in the template, and the resulting
code will not compile unless ICAROUS is present as an app.

Prior commits have removed ICAROUS-specific content from the template and
replaced it with template variables that will be expanded to fit the user's
needs.

This commit modifies the README to explain what those variables are, and how
the data should be passed to Ogma.