ogma-core: Remove ICAROUS-specific list of known variables from cFS backend. Refs #237.
The cFS application template was originally built for ICAROUS, but it is now
generic to cFS. ICAROUS is still mentioned in the template, and the resulting
code will not compile unless ICAROUS is present as an app.

Prior commits have removed ICAROUS-specific content from the template and
replaced it with template variables that will be expanded to fit the user's
needs. However, the backend still assumes that a position variable will be
published by ICAROUS in a cFS message if no variables are passed by the user
when invoking `ogma`.

This commit modifies cFS backend to remove that mention of ICAROUS variables.
It now becomes mandatory for the user to provide some variables if they expect
to be able to monitor data.