ogma-cli: Fix formatting of template variables in README. Refs #222.
The formatting of template variables in the "Template Customization"
section of the F' backend is incorrect. They are being rendered as plain
text, when they should use backticks to be formatted as code.

This commit modifies the template variables in the F' section to use
backticks so that they are properly formatted.