ogma-core: Re-structure cFS backend to avoid nested conditions. Refs #242.
The cFS backend checks whether the arguments are valid and are well
formed, returning different error messages if not, before actually
running. Currently, the implementation of this code results in multiple
nested conditions, which makes the code difficult to maintain and expand
with additional arguments.

This commit re-structures the cFS backend to take advantage of the
ExceptT monad transformer, which stops at the first error. This change
makes the code sequential instead of nested. This implementation does
not alter the behavior of the backend, but it removes some additional
code that could only be triggered if there was an issue in the code,
since it's handled now automatically by the ExceptT transformer.