ogma-core: Update template to use new variable names and structures. Refs #244.
The current ROS backend is overly specific, complex and composes data
manually. For example, it assumes a specific structure for the message
handlers. It also extracts data from data structures, manually converts
it so strings, and then packs it into a JSON structure. This limits
user's abilities to adjust the template to fit their needs, since most
of the data is given as fixed strings, rather than letting users access
each variable declaration and, within them, each type and name
individually.

A prior commit has modified the ROS backend to use provide more raw data
to the ROS template, making it more versatile.

This commit updates the template to make use of such data, by moving
some of the strings that the prior ROS backend was composing together
into the template directly. The template is modified such that it
generates the exact same final ROS application as before.