ogma-core: Make structured data available to FPrime template. Refs #246.
The current FPrime backend is overly specific, complex and composes data
manually. For example, it assumes a specific structure for the message
handlers. It also extracts data from data structures, manually converts
it to strings, and then packs it into a JSON structure. This limits
user's abilities to adjust the template to fit their needs, since most
of the data is given as fixed strings, rather than letting users access
each variable declaration and, within them, each type and name
individually.

This commit simplifies the FPrime backend by packing all the raw data
that may be relevant for the FPrime application in a data structure,
converting it to JSON, and then passing it to the template expansion
system.