ogma-cli: Update README with new FPrime template variables. Refs #246.
The current FPrime backend is overly specific, complex and composes data
manually. For example, it assumes a specific structure for the message
handlers. It also extracts data from data structures, manually converts
it to strings, and then packs it into a JSON structure. This limits
user's abilities to adjust the template to fit their needs, since most
of the data is given as fixed strings, rather than letting users access
each variable declaration and, within them, each type and name
individually.

Prior commits have modified the FPrime backend to provide more raw data
to the FPrime template, and have adjusted the default template
accordingly.

This commit updates the README to document the new variables available
to the template.