ogma-cli: Adjust CLI to match new FPrime backend API. Refs #248.
Currently, the cFS, ROS, F' and standalone backends perform similar
functionality and are built around similar ideas, but are built slightly
differently. This causes an unnecessary amount of code replication, and makes
it hard to factorize code out, to incorporate features from a backend into
others, and to connect all 4 backends to the GUI.

A prior commit has adjusted the FPrime backend, both in how the top-level
function is named, in how arguments are passed, and in the names of the options
to the backend.

This commit adjusts the command-line interface to match the backends's new API.