ogma-core: Adjust tests to match new standalone backend API. Refs #248.
Currently, the cFS, ROS, F' and standalone backends perform similar
functionality and are built around similar ideas, but are built slightly
differently. This causes an unnecessary amount of code replication, and makes
it hard to factorize code out, to incorporate features from a backend into
others, and to connect all 4 backends to the GUI.

A prior commit has adjusted the standalone backend, both in how the top-level
function is named, and in how arguments are passed.

This commit adjusts the tests to match the backends's new API.