ogma-core: Update FPrime backend to process a template variables file. Refs #250.
The generated ROS, FPrime and standalone application may need
definitions that come from external files, such as topic names from
other applications, or structs defined elsewhere. Users may also want to
customize additional aspects of the generated applications without
having to modify the template, such as passing additional variables to
be expanded in the template.

This commit extends the FPrime backend to expect an additional argument
denoting a JSON file with an object whose keys will be expanded as
variables in the FPrime template.