ogma-cli: Expose template vars argument to FPrime backend in CLI. Refs #250.
The generated ROS, FPrime and standalone application may need
definitions that come from external files, such as topic names from
other applications, or structs defined elsewhere. Users may also want to
customize additional aspects of the generated applications without
having to modify the template, such as passing additional variables to
be expanded in the template.

A prior commit has extended the FPrime backend to expect an additional
argument filename containing a JSON object, whose keys are expanded as
variables in the FPrime template.

This commit exposes the argument to the user in the command-line via an
optional CLI argument.