ogma-cli: Update README to document new template vars argument. Refs #250.
The generated ROS, FPrime and standalone application may need
definitions that come from external files, such as topic names from
other applications, or structs defined elsewhere. Users may also want to
customize additional aspects of the generated applications without
having to modify the template, such as passing additional variables to
be expanded in the template.

Prior commits have extended the `ros`, `fprime` and `standalone`
commands to accept an optional argument with a filename containing a
JSON object, whose keys are expanded as variables in the templates.

This commit documents the new argument to the `ros` and `fprime`
commands, and how variables are expanded in application templates.
There is no section in the README for the `standalone` command, so that
remains undocumented.