ogma-core: Make cFS backend accept spec as input. Refs #252.
The cFS backend is lacking in features when compared with other backends
like ROS and F'. In particular, it doesn't accept a specification as
input, which means that the user must take an additional step of
extracting the variable names and handlers from the spec before using
the cfs command.

This commit modifies the cFS backend to accepts a spec as input if
provided, uses it to determine the variables and handlers to use in the
cFS application, accepts selecting the format for the file and the
properties, and can filter properties via an external pre-filter.