ogma-cli: Expose spec processing arguments to cFS backend in CLI. Refs #252.
The cFS backend is lacking in features when compared with other backends
like ROS and F'. In particular, it doesn't accept a specification as
input, which means that the user must take an additional step of
extracting the variable names and handlers from the spec before using
the cfs command.

A prior commit has extended the cFS backend to accept an additional
argument filename containing an input specification, as well as
arguments indicating the format of the input file, the format of the
properties, and a pre-processor for properties.

This commit exposes the arguments to the user in the command-line via
additional CLI flags. With this change, a variables file is no longer
mandatory, so its default value in the CLI is removed to prevent the
backend from always trying to use that file.