ogma-cli: Update README to document new cFS command flags. Refs #252.
The cFS backend is lacking in features when compared with other backends
like ROS and F'. In particular, it doesn't accept a specification as
input, which means that the user must take an additional step of
extracting the variable names and handlers from the spec before using
the cfs command.

Prior commits have extended the cFS backend to accept an additional
argument filename containing an input specification, as well as
arguments indicating the format of the input file, the format of the
properties, and a pre-processor for properties.

This commit documents the new arguments to the `cfs` command in the
README.