ogma-core: Export Spec to monitor logic in standalone backend. Refs #107.
Generating the cFS, ROS or FPrime applications requires multiple steps:
one for the application in the target platform, one for the monitor,
etc. This process is not ideal, and not intuitive from the user's
perspective. Since the spec is already being provided to the backend,
but users still need to call the standalone backend manually to produce
the Copilot spec and copy it to the target directory, in the right
location.

Currently, the standalone backend can only be called via a function that
parses the input specification, and expands the monitoring variables in
a template. This is unnecessarily cumbersome: to make use of the
standalone backend, other backends would have to compose a "fake"
CommandOptions struct, and adjust the template directory manually to
match. This would make the backend aware of the layout of the template,
making the result inflexible.

This commit creates and exports a function that implements the logic of
the standalone backend once a Spec has been parsed, and before the
output has been expanded in the template. This makes it possible for
other backends to call the standalone backend on abstract values and add
the results as part of their own application data before it is expanded
in templates.