ogma-core: Make FPrime backend to produce a Copilot monitor. Refs #107.
Generating the ROS, cFS or FPrime applications requires multiple steps:
one for the application in the target platform, one for the monitor,
etc. This process is not ideal, and not intuitive from the user's
perspective. Since the spec is already being provided to the backend,
but users still need to call the standalone backend manually to produce
the Copilot spec and copy it to the target directory, in the right
location.

A prior commit has modified the standalone backend to provide a function
that processes a spec and calculates the values to expand in the
template.

This commit modifies the FPrime backend to make use of those extensions,
by adding the monitoring data as part of its structure, and using it to
fill in the gaps in a Copilot monitor included with the template.