ogma-core: Make Spec parsing function independent from backend. Refs #107.
Generating the cFS, ROS or FPrime applications requires multiple steps:
one for the application in the target platform, one for the monitor,
etc. This process is not ideal, and not intuitive from the user's
perspective. Since the Spec is already being provided to the backend,
but users still need to call the standalone backend manually to produce
the Copilot spec and copy it to the target directory, in the right
location.

It is easier to have all backends use the standalone backend if they use
the same types for arguments. Currently, code is replicated across
different modules, which makes the types trivially isomorphic, but it
requires data conversion to call one module from another. To make
sharing of that code possible, we need to make the functions that need
to be shared independent of any specific backend.

This commit makes the `parseInputFile` function in each module, which
they use to parse input specifications, independent from the
CommandOptions of that backend. Doing so will make it possible for all
backends to call the same function.