ogma-core: Factorize backend code into shared module. Refs #107.
Generating the cFS, ROS or FPrime applications requires multiple steps:
one for the application in the target platform, one for the monitor,
etc. This process is not ideal, and not intuitive from the user's
perspective. Since the spec is already being provided to the backend,
but users still need to call the standalone backend manually to produce
the Copilot spec and copy it to the target directory, in the right
location.

It is easier to have all backends use the standalone backend if they use
the same types for arguments. Currently, code is replicated across
different modules, which makes the types trivially isomorphic, but it
requires data conversion to call one module from another. Many functions
are also the same in multiple modules.

This commit factorizes code out from each backend into a `Common`
module. Imports are adjusted and reduced accordingly.