ogma-cli: Adjust CI job to install Copilot globally. Refs #107.
Generating the ROS, cFS or FPrime applications requires multiple steps:
one for the application in the target platform, one for the monitor,
etc. This process is not ideal, and not intuitive from the user's
perspective. Since the spec is already being provided to the backend,
but users still need to call the standalone backend manually to produce
the Copilot spec and copy it to the target directory, in the right
location.

Prior commits have extended all backends with the possibility of
leveraging the standalone backend to produce a Copilot spec as part of
the application generated.

To test the new capabilities in a CI job, it is convenient to be able to
run `runhaskell` from any directory, which we can do if we install
Copilot using `cabal install --lib` rather than sandboxes.

This commit modifies the CI to install Copilot as globally available to
the user. To make that easier, we use a newer version of GHC and Cabal.