ogma-cli: Make ROS CI job use spec in test. Refs #107.
Generating the ROS, cFS or FPrime applications requires multiple steps:
one for the application in the target platform, one for the monitor,
etc. This process is not ideal, and not intuitive from the user's
perspective. Since the spec is already being provided to the backend,
but users still need to call the standalone backend manually to produce
the Copilot spec and copy it to the target directory, in the right
location.

Prior commits have extended all backends with the possibility of
leveraging the standalone backend to produce a Copilot spec as part of
the application generated.

This commit leverages those capabilities to generate a full ROS 2
monitoring package during in a CI job and compile the generated code.