ogma-core: Introduce default variable DB data file. Refs #256.
Currently, each backend uses a different format for the variable DBs,
and the format doesn't help users understand what is expected in each
field. This makes the backends too ad hoc, and makes the format hard to
understand.

A prior commit has defined a new type to represent variable DBs, together
with JSON-based parsing functions.

This commit introduces a standard variable DB that contains all the
information specific to ROS and FPrime.