ogma-cli: Update format of variable DB in cFS example. Refs #256.
Currently, each backend uses a different format for the variable DBs,
and the format doesn't help users understand what is expected in each
field. This makes the backends too ad hoc, and makes the format hard to
understand.

Prior commits have defined a new type to represent variable DBs and have
adjusted backends accordingly.

This commit replaces the cFS variable DB included as an example with
Ogma to use the new format.