ogma-cli: Adjust Github Action workflow to use new variable DB format. Refs #256.
Currently, each backend uses a different format for the variable DBs,
and the format doesn't help users understand what is expected in each
field. This makes the backends too ad hoc, and makes the format hard to
understand.

Prior commits have defined a new type to represent variable DBs,
adjusted backends accordingly, and updated the ROS variable DB included
with Ogma as an example.

This commit updates the Github Action to use the new file name and
format during generation of a sample ROS application.