ogma-core: Standardize variable DBs. Refs #256.
Currently, each backend uses a different format for the variable DBs,
and the format doesn't help users understand what is expected in each
field. This makes the backends too ad hoc, and makes the format hard to
understand.

This commit defines a new type to represent variable DBs, as well as
functions to obtain information from and merge variable DBs.