ogma-core: Move ErrorTriplet and ErrorCode to own module. Refs #256.
Currently, each backend uses a different format for the variable DBs,
and the format doesn't help users understand what is expected in each
field. This makes the backends too ad hoc, and makes the format hard to
understand.

In order to introduce a new module that standardizes variable DBs, we
need to make it possible for that module to use error codes so that
errors are reported in the same way as it is done elsewhere in Ogma.

This commit moves ErrorTriplet and ErrorCode to a separate module, and
adjusts exports and imports across modules accordingly.