ogma-core: Provide parsing functions for the new variable DB format. Refs #256.
Currently, each backend uses a different format for the variable DBs,
and the format doesn't help users understand what is expected in each
field. This makes the backends too ad hoc, and makes the format hard to
understand.

Prior commits have introduced a new type to represent variable DBs,
functions to read variable DBs and to obtain information from them.

This commit replaces the parsing function and other definitions in the
module to use the new format.