ogma-cli: Update README to new variable DB format. Refs #256.
Currently, each backend uses a different format for the variable DBs,
and the format doesn't help users understand what is expected in each
field. This makes the backends too ad hoc, and makes the format hard to
understand.

Prior commits have defined a new type to represent variable DBs,
adjusted backends accordingly, and updated the cFS and ROS variable DBs
included with Ogma as examples.

This commit updates the instructions included in the README to refer to
the example files by their new names, and to use the new file format
where applicable.