ogma-extra: Add auxiliary functions to manipulate strings, lists. Refs #256.
Currently, each backend uses a different format for the variable DBs,
and the format doesn't help users understand what is expected in each
field. This makes the backends too ad hoc, and makes the format hard to
understand.

This commit adds auxiliary functions to convert strings to PascalCase,
and to drop a suffix from a list, so that they can be used by the
backends to produce useful names for data processing functions.