ogma-language-jsonspec: Extend JSONSpec with additional data associated with results. Refs #219.
Currently, the monitor notifications to the rest of the system happen in
one of two ways: as a call to a user-defined function, or as a
notification in the message bus or via a port. None of those messages
carry any additional information: the occurrence of a property or
requirement violation is implicit in the fact that the notification is
present.

This commit modifies the definition of a JSON spec so that additional data
can be associated to a requirement result, namely a type for the data,
and an expression.