ogma-core: Adjust function to obtain handlers to include additional data. Refs #219.
Currently, the monitor notifications to the rest of the system happen in
one of two ways: as a call to a user-defined function, or as a
notification in the message bus or via a port. None of those messages
carry any additional information: the occurrence of a property or
requirement violation is implicit in the fact that the notification is
present.

Prior commits have modified several spec definitions to that additional
data (i.e., a type and an expression) can be associated with a
requirement. In order to use that information in the backends, we need
to provide a mechanism to find the type associated with a notification
type (e.g., message, port) even if no input variable is associated to
such type.

This commit modifies the function that provides the handlers or monitors
so that it includes the type of any additional data that may be
associated to such monitors or handlers.