ogma-core: Make Copilot spec pass arguments to triggers where applicable. Refs #219.
Currently, the monitor notifications to the rest of the system happen in
one of two ways: as a call to a user-defined function, or as a
notification in the message bus or via a port. None of those messages
carry any additional information: the occurrence of a property or
requirement violation is implicit in the fact that the notification is
present.

A prior commit has modified the definition of a Spec so that additional
data (i.e., a type and an expression) can be associated with a
requirement.

This commit makes use of that data in the Copilot generator so that
handlers use the additional data when available.